allow protocols with executables from `$PATH`
When a protocol or unmocked command specifies an executable without
a path (e.g. `cp` instead of `/bin/cp`) check-protocols will refer to
the executable location using the `$PATH` variable when comparing
against commands executed by a script. When displaying commands
executables that are in a location in `$PATH` will be automatically
shortened to exclude the path (e.g. an error message will display `cp
a b` instead of `/bin/cp a b`).