better errors in case the scriptkeeper_executable can't be found
This can happen during tests when running `cargo test --lib` when
the scriptkeeper executable wasn't built before. The proper fix
would be to actually make sure we build the executable before
running these tests, but I'm not sure that's possible. In any
case this error message makes it easier to track down.